requirements.txt
# TELBOT - Ultimate Modular LLM Telegram Bot Requirements
# Optimized for Ubuntu 22.04, 8GB RAM, 4 CPU Cores - May 2025

# ================================
# CORE TELEGRAM BOT FRAMEWORK
# ================================
python-telegram-bot==21.1.1
telegram==0.0.1

# ================================
# LOCAL LLM SUPPORT (CPU-OPTIMIZED)
# ================================
llama-cpp-python==0.2.78
transformers==4.40.2
torch==2.3.0+cpu --index-url https://download.pytorch.org/whl/cpu
tokenizers==0.19.1
accelerate==0.30.1
safetensors==0.4.3

# ================================
# LANGCHAIN ECOSYSTEM
# ================================
langchain==0.2.1
langchain-community==0.2.1
langchain-core==0.2.1
langchain-experimental==0.0.59
langchain-huggingface==0.0.3

# ================================
# HUGGING FACE INTEGRATION
# ================================
huggingface-hub==0.23.0
datasets==2.19.1
sentence-transformers==2.7.0

# ================================
# DATA PROCESSING & DATABASE
# ================================
pandas==2.2.2
numpy==1.26.4
sqlalchemy==2.0.30
sqlite3worker==1.1.7
openpyxl==3.1.2
xlrd==2.0.1

# ================================
# HTTP & API UTILITIES
# ================================
requests==2.31.0
aiohttp==3.9.5
httpx==0.27.0
urllib3==2.2.1

# ================================
# ENVIRONMENT & CONFIGURATION
# ================================
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
dynaconf==3.2.5

# ================================
# TEXT PROCESSING & NLP
# ================================
nltk==3.8.1
spacy==3.7.4
regex==2024.5.15
beautifulsoup4==4.12.3
markdown==3.6
textstat==0.7.3

# ================================
# ASYNC SUPPORT
# ================================
asyncio==3.4.3
aiofiles==23.2.0
asyncpg==0.29.0

# ================================
# LOGGING & MONITORING
# ================================
loguru==0.7.2
rich==13.7.1
prometheus-client==0.20.0
structlog==24.1.0

# ================================
# SECURITY & VALIDATION
# ================================
cryptography==42.0.7
validators==0.28.1
passlib==1.7.4
bcrypt==4.1.2

# ================================
# CACHING & PERFORMANCE
# ================================
redis==5.0.4
diskcache==5.6.3
cachetools==5.3.3

# ================================
# FILE HANDLING & I/O
# ================================
pathlib2==2.3.7
chardet==5.2.0
python-magic==0.4.27
pillow==10.3.0

# ================================
# DATA VISUALIZATION (OPTIONAL)
# ================================
plotly==5.22.0
matplotlib==3.8.4
seaborn==0.13.2

# ================================
# SYSTEM MONITORING
# ================================
psutil==5.9.8
memory-profiler==0.61.0
py-cpuinfo==9.0.0
gputil==1.4.0

# ================================
# TASK QUEUE & SCHEDULING
# ================================
celery==5.3.7
schedule==1.2.2
apscheduler==3.10.4

# ================================
# DEVELOPMENT & TESTING
# ================================
pytest==8.2.1
pytest-asyncio==0.23.6
pytest-mock==3.12.0
pytest-cov==5.0.0
black==24.4.2
flake8==7.0.0
mypy==1.10.0
pre-commit==3.7.1

# ================================
# DATABASE EXTRAS
# ================================
alembic==1.13.1
psycopg2-binary==2.9.9

# ================================
# DOCUMENT PROCESSING
# ================================
pypdf==4.2.0
python-docx==1.1.0
python-pptx==0.6.23

# ================================
# WEB FRAMEWORK (FOR API)
# ================================
fastapi==0.111.0
uvicorn==0.29.0
starlette==0.37.2

# ================================
# MACHINE LEARNING EXTRAS
# ================================
scikit-learn==1.4.2
scipy==1.13.0
joblib==1.4.0

# ================================
# UTILITY LIBRARIES
# ================================
python-dateutil==2.9.0
pytz==2024.1
click==8.1.7
tqdm==4.66.4
jinja2==3.1.4
pyyaml==6.0.1
toml==0.10.2

# ================================
# CLOUD & API INTEGRATIONS
# ================================
boto3==1.34.94
google-cloud-storage==2.16.0
azure-storage-blob==12.19.1

# ================================
# BACKUP & COMPRESSION
# ================================
zipfile36==0.1.3
tarfile==0.11.0

# ================================
# MODEL-SPECIFIC PROVIDERS
# ================================
openai==1.30.1
anthropic==0.25.7
cohere==5.2.4

# ================================
# SPECIALIZED TOOLS
# ================================
streamlit==1.34.0  # For admin dashboard
gradio==4.26.0     # For model testing interface
jupyter==1.0.0     # For development notebooks

# ================================
# OLLAMA INTEGRATION
# ================================
ollama==0.1.9

# ================================
# AUDIO/VIDEO PROCESSING (OPTIONAL)
# ================================
librosa==0.10.1
opencv-python==4.9.0.80

# ================================
# GRAPH & NETWORK ANALYSIS
# ================================
networkx==3.3
pygraphviz==1.11

# ================================
# TIMEZONE & LOCALIZATION
# ================================
babel==2.14.0
pytz-deprecation-shim==0.1.0

# ================================
# PROFILING & DEBUGGING
# ================================
line-profiler==4.1.1
py-spy==0.3.14